Rearranged some code to avoid cyclic imports
