+++ tests/api/test_helpers.py
@@ -125,7 +125,7 @@ def test_reject_without_file():
 
 
 def custom_action(structure, options):
-    return (structure, options)
+    return structure, options
 
 
 custom_action.__module__ = 'awesome_module'
@@ -133,11 +133,7 @@ custom_action.__module__ = 'awesome_module'
 
 def init_git(structure, options):
     """Fake action that shares the same name as a default action."""
-    return (structure, options)
-
-
-def test_get_id():
-    assert helpers.get_id(custom_action) == 'awesome_module:custom_action'
+    return structure, options
 
 
 def test_register_before():
